refactorized vector in matrix class, now takes doubles

+++ headers/matrix.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include <vector>
 #include <string>
+#include <vector>
 
 class Matrix {
 public:
@@ -13,11 +13,11 @@ public:
     bool isInputValuesEqualToColumnsSize(std::string temporaryRow);
     bool isInputValuesConsistsDigits(std::string temporaryRow);
 
-    std::vector<int> createVectorOfIntsFromTemporaryRowInput(std::string temporaryRow);
+    std::vector<double> createVectorOfIntsFromTemporaryRowInput(std::string temporaryRow);
 
     void printMatrix();
 
-    std::vector<std::vector<int>>* getMatrix() const
+    std::vector<std::vector<double>>* getMatrix() const
     {
         return matrix_;
     }
@@ -32,7 +32,7 @@ public:
     }
 
 private:
-    std::vector<std::vector<int>>* matrix_{};
+    std::vector<std::vector<double>>* matrix_{};
     std::size_t rows_{};
     std::size_t columns_{};
 };
